added embedValue() to Function
It takes a value or pointer and embeds it in the generated code. The return
value is a wrapper type which was initialized with this value. It is mutable
and can be changed like any other wrapper type.

It is especially useful to embed pointers to data structures of the host which
the generated code can read and manipulate through the pointer. Of course, the
data has to be still valid when generated function tries to access it. For
things with varying memory locations, better pass the pointer as argument to
the generated function.