ported LLVM to version 11, only sumLLVM example ported, yet
The LLVM submodule in external was updated to version 11.0. The shell script
buildDependencies.sh was changed slightly to enable perf profiling support in
LLVM. Improved profiling and debugging support are the main reasons for
upgrading LLVM.

runtimellvmjit.h and llvmjit/Function.h were the main files which needed to be
ported. Porting is not done yet, though. The example sumLLVM is the only one
working at the moment.