use environment variables to enable code dumping
Dumping of profiling and debugging information is disabled by default in the
LLVM backend. Profiling with perf requires dumping of the object file in the
jitdump format. This is now only enabled when COAT_JITDUMP=1 is set.

Similarly, for debugging purpose, the generated code can be dumped as normal
object file. This is enabled by COAT_OBJDUMP=1 and writes the file in the
current working directory.