added debug info to Struct types
The members of a struct are accessible and have the correct data types, but
the name of the struct and the names of the members are currently hardcoded.
They are not the same as the identifiers used in the C++ code. More macros
for annotation are needed to preserve the names for use during runtime.